test/main/java/io/FileInputTest.java

package main.java.io;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

public class FileInputTest {
    
    public FileInputTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    @Test
    public void testReadNext() {
        FileInput in = new FileInput(new MockInStream("te"));
        
        assertEquals(116, in.readNext());
        assertEquals(101, in.readNext());
    }
}
